+++ cocos/2d/platform/CCEGLViewProtocol.cpp
@@ -37,7 +37,7 @@ namespace {
     static Touch* g_touches[EventTouch::MAX_TOUCHES] = { nullptr };
     static unsigned int g_indexBitsUsed = 0;
     // System touch pointer ID (It may not be ascending order number) <-> Ascending order number from 0
-    static std::map<long, int> g_touchIdReorderMap;
+    static std::map<int, int> g_touchIdReorderMap;
     
     static int getUnUsedIndex()
     {
@@ -223,9 +223,9 @@ const std::string& EGLViewProtocol::getViewName() const
     return _viewName;
 }
 
-void EGLViewProtocol::handleTouchesBegin(int num, long ids[], float xs[], float ys[])
+void EGLViewProtocol::handleTouchesBegin(int num, int ids[], float xs[], float ys[])
 {
-    long id = 0;
+    int id = 0;
     float x = 0.0f;
     float y = 0.0f;
     int unusedIndex = 0;
@@ -273,9 +273,9 @@ void EGLViewProtocol::handleTouchesBegin(int num, long ids[], float xs[], float 
     dispatcher->dispatchEvent(&touchEvent);
 }
 
-void EGLViewProtocol::handleTouchesMove(int num, long ids[], float xs[], float ys[])
+void EGLViewProtocol::handleTouchesMove(int num, int ids[], float xs[], float ys[])
 {
-    long id = 0;
+    int id = 0;
     float x = 0.0f;
     float y = 0.0f;
     EventTouch touchEvent;
@@ -305,7 +305,7 @@ void EGLViewProtocol::handleTouchesMove(int num, long ids[], float xs[], float y
         else
         {
             // It is error, should return.
-            CCLOG("Moving touches with id: %ld error", id);
+            CCLOG("Moving touches with id: %d error", id);
             return;
         }
     }
@@ -321,9 +321,9 @@ void EGLViewProtocol::handleTouchesMove(int num, long ids[], float xs[], float y
     dispatcher->dispatchEvent(&touchEvent);
 }
 
-void EGLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode, int num, long ids[], float xs[], float ys[])
+void EGLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode, int num, int ids[], float xs[], float ys[])
 {
-    long id = 0;
+    int id = 0;
     float x = 0.0f;
     float y = 0.0f;
     EventTouch touchEvent;
@@ -358,7 +358,7 @@ void EGLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode
         } 
         else
         {
-            CCLOG("Ending touches with id: %ld error", id);
+            CCLOG("Ending touches with id: %d error", id);
             return;
         } 
 
@@ -381,12 +381,12 @@ void EGLViewProtocol::handleTouchesOfEndOrCancel(EventTouch::EventCode eventCode
     }
 }
 
-void EGLViewProtocol::handleTouchesEnd(int num, long ids[], float xs[], float ys[])
+void EGLViewProtocol::handleTouchesEnd(int num, int ids[], float xs[], float ys[])
 {
     handleTouchesOfEndOrCancel(EventTouch::EventCode::ENDED, num, ids, xs, ys);
 }
 
-void EGLViewProtocol::handleTouchesCancel(int num, long ids[], float xs[], float ys[])
+void EGLViewProtocol::handleTouchesCancel(int num, int ids[], float xs[], float ys[])
 {
     handleTouchesOfEndOrCancel(EventTouch::EventCode::CANCELLED, num, ids, xs, ys);
 }
